[HAPI-182] Blend the old models with the DB class models
## [HAPI-182] Description
Blending the old models to work only with the DB class models
