+++ src/controllers/auth/send-email-confirm.js
@@ -1,4 +1,4 @@
-const User = require('4hands-api/src/models/collections/User');
+const User = require('4hands-api/src/collections/Models/users.model');
 const Endpoint = require('4hands-api/src/models/settings/Endpoint');
 const AuthService = require('4hands-api/src/services/Auth');
