+++ src/models/index.js
@@ -1,7 +1,5 @@
 const routes = require('./routes');
-const User = require('./collections/User');
 
 module.exports = {
-    routes,
-    User
+    routes
 };
